Move image processing code to processing directory
