+++ processing/gearcontours.py
