+++ processing/tapecontours.py
